Added overview doc, updated imgs, minor changes
